refactor: externalize play furniture layout
