+++ src/systems/furnitureLayout.ts
@@ -1,0 +1,247 @@
+import type { FurnitureLayoutEntry } from './SearchSystem';
+
+export const BEDROOM_FURNITURE_LAYOUT: FurnitureLayoutEntry[] = [
+  {
+    x: 330,
+    y: 200,
+    options: {
+      searchable: true,
+      name: 'Left Upper Bed',
+      searchDuration: 2600,
+      checkPoints: [0.85, 0.55, 0.25],
+      findChance: 0.5,
+      emoji: '🛏️',
+      sprite: {
+        frame: 'bed',
+        offsetY: -20,
+        depth: 2,
+        scale: 0.7,
+      },
+      hitbox: {
+        width: 176,
+        height: 224,
+        offsetX: -1.5,
+        offsetY: 0,
+      },
+    },
+  },
+  {
+    x: 640,
+    y: 200,
+    options: {
+      searchable: true,
+      name: 'Middle Upper Bed',
+      searchDuration: 2600,
+      checkPoints: [0.85, 0.55, 0.25],
+      findChance: 0.5,
+      emoji: '🛏️',
+      sprite: {
+        frame: 'bed',
+        offsetY: -20,
+        depth: 2,
+        scale: 0.7,
+      },
+      hitbox: {
+        width: 176,
+        height: 224,
+        offsetX: -1.5,
+        offsetY: 0,
+      },
+    },
+  },
+  {
+    x: 950,
+    y: 200,
+    options: {
+      searchable: true,
+      name: 'Right Upper Bed',
+      searchDuration: 2600,
+      checkPoints: [0.85, 0.55, 0.25],
+      findChance: 0.5,
+      emoji: '🛏️',
+      sprite: {
+        frame: 'bed',
+        offsetY: -20,
+        depth: 2,
+        scale: 0.7,
+        flipX: true,
+      },
+      hitbox: {
+        width: 176,
+        height: 224,
+        offsetX: -1.5,
+        offsetY: 0,
+      },
+    },
+  },
+  {
+    x: 950,
+    y: 520,
+    options: {
+      searchable: true,
+      name: 'Right Lower Bed',
+      searchDuration: 2600,
+      checkPoints: [0.85, 0.55, 0.25],
+      findChance: 0.5,
+      emoji: '🛏️',
+      sprite: {
+        frame: 'bed',
+        offsetY: -20,
+        depth: 2,
+        scale: 0.7,
+        flipX: true,
+      },
+      hitbox: {
+        width: 176,
+        height: 224,
+        offsetX: -1.5,
+        offsetY: 0,
+      },
+    },
+  },
+  {
+    x: 150,
+    y: 200,
+    options: {
+      searchable: true,
+      name: 'Locker A',
+      searchDuration: 2200,
+      checkPoints: [0.75, 0.4],
+      findChance: 0.6,
+      emoji: '🎒',
+      sprite: {
+        frame: 'dresser',
+        offsetY: -40,
+        depth: 2,
+        scale: 0.7,
+      },
+      hitbox: {
+        width: 95,
+        height: 153,
+        offsetX: 1,
+        offsetY: 0.5,
+      },
+    },
+  },
+  {
+    x: 150,
+    y: 520,
+    options: {
+      searchable: true,
+      name: 'Locker B',
+      searchDuration: 2200,
+      checkPoints: [0.75, 0.4],
+      findChance: 0.6,
+      emoji: '🎒',
+      sprite: {
+        frame: 'dresser',
+        offsetY: -40,
+        depth: 2,
+        scale: 0.7,
+      },
+      hitbox: {
+        width: 95,
+        height: 153,
+        offsetX: 1,
+        offsetY: 0.5,
+      },
+    },
+  },
+  {
+    x: 1110,
+    y: 200,
+    options: {
+      searchable: true,
+      name: 'Locker C',
+      searchDuration: 2200,
+      checkPoints: [0.75, 0.4],
+      findChance: 0.6,
+      emoji: '🎒',
+      sprite: {
+        frame: 'dresser',
+        offsetY: -40,
+        depth: 2,
+        scale: 0.7,
+        flipX: true,
+      },
+      hitbox: {
+        width: 95,
+        height: 153,
+        offsetX: 1,
+        offsetY: 0.5,
+      },
+    },
+  },
+  {
+    x: 1110,
+    y: 520,
+    options: {
+      searchable: true,
+      name: 'Locker D',
+      searchDuration: 2200,
+      checkPoints: [0.75, 0.4],
+      findChance: 0.6,
+      emoji: '🎒',
+      sprite: {
+        frame: 'dresser',
+        offsetY: -40,
+        depth: 2,
+        scale: 0.7,
+        flipX: true,
+      },
+      hitbox: {
+        width: 95,
+        height: 153,
+        offsetX: 1,
+        offsetY: 0.5,
+      },
+    },
+  },
+  {
+    x: 480,
+    y: 650,
+    options: {
+      searchable: true,
+      name: 'Study Desk A',
+      searchDuration: 2200,
+      checkPoints: [0.75, 0.4],
+      findChance: 0.6,
+      emoji: '🖥️',
+      sprite: {
+        frame: 'desk',
+        offsetY: -30,
+        depth: 2,
+        scale: 0.85,
+      },
+      hitbox: {
+        width: 204,
+        height: 123,
+        offsetX: 1.5,
+      },
+    },
+  },
+  {
+    x: 800,
+    y: 650,
+    options: {
+      searchable: true,
+      name: 'Study Desk B',
+      searchDuration: 2200,
+      checkPoints: [0.75, 0.4],
+      findChance: 0.6,
+      emoji: '🖥️',
+      sprite: {
+        frame: 'desk',
+        offsetY: -30,
+        depth: 2,
+        scale: 0.85,
+        flipX: true,
+      },
+      hitbox: {
+        width: 204,
+        height: 123,
+        offsetX: 1.5,
+      },
+    },
+  },
+];
